requirements.txt
praw==7.7.1
asyncpraw==7.8.1
whoosh==2.7.4
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.26.0
sqlalchemy==2.0.27
aiohttp==3.9.3
python-multipart==0.0.9
jinja2==3.1.3
chromadb==0.4.24
sentence-transformers==2.5.1
beautifulsoup4==4.12.3
pyngrok==6.0.0 